word2vec changed type to signed

+++ tasks/word2vec/word2vec.cpp
@@ -2,8 +2,8 @@
 #include <string>
 #include <vector>
 
-size_t CountProduct(const std::vector<int>& a, const std::vector<int>& b) {
-    size_t res = 0;
+int64_t CountProduct(const std::vector<int>& a, const std::vector<int>& b) {
+    int64_t res = 0;
     for (size_t i = 0; i < a.size(); ++i) {
         res += a[i] * b[i];
     }
@@ -16,9 +16,9 @@ std::vector<std::string> FindClosestWords(const std::vector<std::string>& words,
         return {};
     }
     std::vector<std::string> ans;
-    size_t mx = CountProduct(vectors[0], vectors[1]);
+    int64_t mx = CountProduct(vectors[0], vectors[1]);
     for (size_t i = 2; i < words.size(); ++i) {
-        size_t cnt = CountProduct(vectors[0], vectors[i]);
+        int64_t cnt = CountProduct(vectors[0], vectors[i]);
         if (cnt > mx) {
             mx = cnt;
             ans = {words[i]};
